Initial genesis file (#1)
# Background

So that we have a history of genesis file parameter changes before we go
to mainnet so that we don't forget them.

This is a genesis that I've taken from testnet-10 and deleted all
accounts. We will need to run gentxs for us to add first validator in,
but that's something that we will do before going live :).